corrections

+++ maggie_motor_controller/src/motor_controller_node.cpp
@@ -363,7 +363,7 @@ bool MotorControllerNode::joint_calibration(maggie_motor_controller_msgs::MoveAb
 
 		_driver->calibrateDriver(int( _calibration_home * factor));
 		
-		ROS_INFO("Home in pulses: %f\tFactor: %d\tHome in radians: %f\t", _calibration_home * factor, factor, _calibration_home);		
+		ROS_INFO("Home in pulses: %d\tFactor: %d\tHome in radians: %f\t", int(_calibration_home * factor), factor, _calibration_home);		
 
 		return true;
 }
